Authentication transport: Split into own files
