+++ cloud/jira_test.go
@@ -450,207 +450,3 @@ func TestClient_GetBaseURL_WithURL(t *testing.T) {
 		t.Errorf("Base URLs are not equal. Expected %+v, got %+v", *u, b)
 	}
 }
-
-func TestBasicAuthTransport(t *testing.T) {
-	setup()
-	defer teardown()
-
-	username, password := "username", "password"
-
-	testMux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		u, p, ok := r.BasicAuth()
-		if !ok {
-			t.Errorf("request does not contain basic auth credentials")
-		}
-		if u != username {
-			t.Errorf("request contained basic auth username %q, want %q", u, username)
-		}
-		if p != password {
-			t.Errorf("request contained basic auth password %q, want %q", p, password)
-		}
-	})
-
-	tp := &BasicAuthTransport{
-		Username: username,
-		Password: password,
-	}
-
-	basicAuthClient, _ := NewClient(tp.Client(), testServer.URL)
-	req, _ := basicAuthClient.NewRequest("GET", ".", nil)
-	basicAuthClient.Do(req, nil)
-}
-
-func TestBasicAuthTransport_transport(t *testing.T) {
-	// default transport
-	tp := &BasicAuthTransport{}
-	if tp.transport() != http.DefaultTransport {
-		t.Errorf("Expected http.DefaultTransport to be used.")
-	}
-
-	// custom transport
-	tp = &BasicAuthTransport{
-		Transport: &http.Transport{},
-	}
-	if tp.transport() == http.DefaultTransport {
-		t.Errorf("Expected custom transport to be used.")
-	}
-}
-
-// Test that the cookie in the transport is the cookie returned in the header
-func TestCookieAuthTransport_SessionObject_Exists(t *testing.T) {
-	setup()
-	defer teardown()
-
-	testCookie := &http.Cookie{Name: "test", Value: "test"}
-
-	testMux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		cookies := r.Cookies()
-
-		if len(cookies) < 1 {
-			t.Errorf("No cookies set")
-		}
-
-		if cookies[0].Name != testCookie.Name {
-			t.Errorf("Cookie names don't match, expected %v, got %v", testCookie.Name, cookies[0].Name)
-		}
-
-		if cookies[0].Value != testCookie.Value {
-			t.Errorf("Cookie values don't match, expected %v, got %v", testCookie.Value, cookies[0].Value)
-		}
-	})
-
-	tp := &CookieAuthTransport{
-		Username:      "username",
-		Password:      "password",
-		AuthURL:       "https://some.jira.com/rest/auth/1/session",
-		SessionObject: []*http.Cookie{testCookie},
-	}
-
-	basicAuthClient, _ := NewClient(tp.Client(), testServer.URL)
-	req, _ := basicAuthClient.NewRequest("GET", ".", nil)
-	basicAuthClient.Do(req, nil)
-}
-
-// Test that an empty cookie in the transport is not returned in the header
-func TestCookieAuthTransport_SessionObject_ExistsWithEmptyCookie(t *testing.T) {
-	setup()
-	defer teardown()
-
-	emptyCookie := &http.Cookie{Name: "empty_cookie", Value: ""}
-	testCookie := &http.Cookie{Name: "test", Value: "test"}
-
-	testMux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		cookies := r.Cookies()
-
-		if len(cookies) > 1 {
-			t.Errorf("The empty cookie should not have been added")
-		}
-
-		if cookies[0].Name != testCookie.Name {
-			t.Errorf("Cookie names don't match, expected %v, got %v", testCookie.Name, cookies[0].Name)
-		}
-
-		if cookies[0].Value != testCookie.Value {
-			t.Errorf("Cookie values don't match, expected %v, got %v", testCookie.Value, cookies[0].Value)
-		}
-	})
-
-	tp := &CookieAuthTransport{
-		Username:      "username",
-		Password:      "password",
-		AuthURL:       "https://some.jira.com/rest/auth/1/session",
-		SessionObject: []*http.Cookie{emptyCookie, testCookie},
-	}
-
-	basicAuthClient, _ := NewClient(tp.Client(), testServer.URL)
-	req, _ := basicAuthClient.NewRequest("GET", ".", nil)
-	basicAuthClient.Do(req, nil)
-}
-
-// Test that if no cookie is in the transport, it checks for a cookie
-func TestCookieAuthTransport_SessionObject_DoesNotExist(t *testing.T) {
-	setup()
-	defer teardown()
-
-	testCookie := &http.Cookie{Name: "does_not_exist", Value: "does_not_exist"}
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		http.SetCookie(w, testCookie)
-		w.Write([]byte(`OK`))
-	}))
-	defer ts.Close()
-
-	testMux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		cookies := r.Cookies()
-
-		if len(cookies) < 1 {
-			t.Errorf("No cookies set")
-		}
-
-		if cookies[0].Name != testCookie.Name {
-			t.Errorf("Cookie names don't match, expected %v, got %v", testCookie.Name, cookies[0].Name)
-		}
-
-		if cookies[0].Value != testCookie.Value {
-			t.Errorf("Cookie values don't match, expected %v, got %v", testCookie.Value, cookies[0].Value)
-		}
-	})
-
-	tp := &CookieAuthTransport{
-		Username: "username",
-		Password: "password",
-		AuthURL:  ts.URL,
-	}
-
-	basicAuthClient, _ := NewClient(tp.Client(), testServer.URL)
-	req, _ := basicAuthClient.NewRequest("GET", ".", nil)
-	basicAuthClient.Do(req, nil)
-}
-
-func TestJWTAuthTransport_HeaderContainsJWT(t *testing.T) {
-	setup()
-	defer teardown()
-
-	sharedSecret := []byte("ssshh,it's a secret")
-	issuer := "add-on.key"
-
-	jwtTransport := &JWTAuthTransport{
-		Secret: sharedSecret,
-		Issuer: issuer,
-	}
-
-	testMux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		// look for the presence of the JWT in the header
-		val := r.Header.Get("Authorization")
-		if !strings.Contains(val, "JWT ") {
-			t.Errorf("request does not contain JWT in the Auth header")
-		}
-	})
-
-	jwtClient, _ := NewClient(jwtTransport.Client(), testServer.URL)
-	jwtClient.Issue.Get("TEST-1", nil)
-}
-
-func TestPATAuthTransport_HeaderContainsAuth(t *testing.T) {
-	setup()
-	defer teardown()
-
-	token := "shhh, it's a token"
-
-	patTransport := &PATAuthTransport{
-		Token: token,
-	}
-
-	testMux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		val := r.Header.Get("Authorization")
-		expected := "Bearer " + token
-		if val != expected {
-			t.Errorf("request does not contain bearer token in the Authorization header.")
-		}
-	})
-
-	client, _ := NewClient(patTransport.Client(), testServer.URL)
-	client.User.GetSelf()
-
-}
