+++ onpremise/jira.go
@@ -3,19 +3,14 @@ package onpremise
 import (
 	"bytes"
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
 	"net/url"
 	"reflect"
-	"sort"
 	"strings"
-	"time"
 
-	jwt "github.com/golang-jwt/jwt/v4"
 	"github.com/google/go-querystring/query"
 )
 
@@ -347,295 +342,4 @@ func (r *Response) populatePageValues(v interface{}) {
 		r.MaxResults = value.MaxResults
 		r.Total = value.Total
 	}
-}
-
-// BasicAuthTransport is an http.RoundTripper that authenticates all requests
-// using HTTP Basic Authentication with the provided username and password.
-type BasicAuthTransport struct {
-	Username string
-	Password string
-
-	// Transport is the underlying HTTP transport to use when making requests.
-	// It will default to http.DefaultTransport if nil.
-	Transport http.RoundTripper
-}
-
-// RoundTrip implements the RoundTripper interface.  We just add the
-// basic auth and return the RoundTripper for this transport type.
-func (t *BasicAuthTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	req2 := cloneRequest(req) // per RoundTripper contract
-
-	req2.SetBasicAuth(t.Username, t.Password)
-	return t.transport().RoundTrip(req2)
-}
-
-// Client returns an *http.Client that makes requests that are authenticated
-// using HTTP Basic Authentication.  This is a nice little bit of sugar
-// so we can just get the client instead of creating the client in the calling code.
-// If it's necessary to send more information on client init, the calling code can
-// always skip this and set the transport itself.
-func (t *BasicAuthTransport) Client() *http.Client {
-	return &http.Client{Transport: t}
-}
-
-func (t *BasicAuthTransport) transport() http.RoundTripper {
-	if t.Transport != nil {
-		return t.Transport
-	}
-	return http.DefaultTransport
-}
-
-// BearerAuthTransport is a http.RoundTripper that authenticates all requests
-// using Jira's bearer (oauth 2.0 (3lo)) based authentication.
-type BearerAuthTransport struct {
-	Token string
-
-	// Transport is the underlying HTTP transport to use when making requests.
-	// It will default to http.DefaultTransport if nil.
-	Transport http.RoundTripper
-}
-
-// RoundTrip implements the RoundTripper interface.  We just add the
-// bearer token and return the RoundTripper for this transport type.
-func (t *BearerAuthTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	req2 := cloneRequest(req) // per RoundTripper contract
-
-	req2.Header.Set("Authorization", fmt.Sprintf("Bearer %s", t.Token))
-	return t.transport().RoundTrip(req2)
-}
-
-// Client returns an *http.Client that makes requests that are authenticated
-// using HTTP Basic Authentication.  This is a nice little bit of sugar
-// so we can just get the client instead of creating the client in the calling code.
-// If it's necessary to send more information on client init, the calling code can
-// always skip this and set the transport itself.
-func (t *BearerAuthTransport) Client() *http.Client {
-	return &http.Client{Transport: t}
-}
-
-func (t *BearerAuthTransport) transport() http.RoundTripper {
-	if t.Transport != nil {
-		return t.Transport
-	}
-	return http.DefaultTransport
-}
-
-// PATAuthTransport is an http.RoundTripper that authenticates all requests
-// using the Personal Access Token specified.
-// See here for more info: https://confluence.atlassian.com/enterprise/using-personal-access-tokens-1026032365.html
-type PATAuthTransport struct {
-	// Token is the key that was provided by Jira when creating the Personal Access Token.
-	Token string
-
-	// Transport is the underlying HTTP transport to use when making requests.
-	// It will default to http.DefaultTransport if nil.
-	Transport http.RoundTripper
-}
-
-// RoundTrip implements the RoundTripper interface.  We just add the
-// basic auth and return the RoundTripper for this transport type.
-func (t *PATAuthTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	req2 := cloneRequest(req) // per RoundTripper contract
-	req2.Header.Set("Authorization", "Bearer "+t.Token)
-	return t.transport().RoundTrip(req2)
-}
-
-// Client returns an *http.Client that makes requests that are authenticated
-// using HTTP Basic Authentication.  This is a nice little bit of sugar
-// so we can just get the client instead of creating the client in the calling code.
-// If it's necessary to send more information on client init, the calling code can
-// always skip this and set the transport itself.
-func (t *PATAuthTransport) Client() *http.Client {
-	return &http.Client{Transport: t}
-}
-
-func (t *PATAuthTransport) transport() http.RoundTripper {
-	if t.Transport != nil {
-		return t.Transport
-	}
-	return http.DefaultTransport
-}
-
-// CookieAuthTransport is an http.RoundTripper that authenticates all requests
-// using Jira's cookie-based authentication.
-//
-// Note that it is generally preferable to use HTTP BASIC authentication with the REST API.
-// However, this resource may be used to mimic the behaviour of Jira's log-in page (e.g. to display log-in errors to a user).
-//
-// Jira API docs: https://docs.atlassian.com/jira/REST/latest/#auth/1/session
-type CookieAuthTransport struct {
-	Username string
-	Password string
-	AuthURL  string
-
-	// SessionObject is the authenticated cookie string.s
-	// It's passed in each call to prove the client is authenticated.
-	SessionObject []*http.Cookie
-
-	// Transport is the underlying HTTP transport to use when making requests.
-	// It will default to http.DefaultTransport if nil.
-	Transport http.RoundTripper
-}
-
-// RoundTrip adds the session object to the request.
-func (t *CookieAuthTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	if t.SessionObject == nil {
-		err := t.setSessionObject()
-		if err != nil {
-			return nil, fmt.Errorf("cookieauth: no session object has been set: %w", err)
-		}
-	}
-
-	req2 := cloneRequest(req) // per RoundTripper contract
-	for _, cookie := range t.SessionObject {
-		// Don't add an empty value cookie to the request
-		if cookie.Value != "" {
-			req2.AddCookie(cookie)
-		}
-	}
-
-	return t.transport().RoundTrip(req2)
-}
-
-// Client returns an *http.Client that makes requests that are authenticated
-// using cookie authentication
-func (t *CookieAuthTransport) Client() *http.Client {
-	return &http.Client{Transport: t}
-}
-
-// setSessionObject attempts to authenticate the user and set
-// the session object (e.g. cookie)
-func (t *CookieAuthTransport) setSessionObject() error {
-	req, err := t.buildAuthRequest()
-	if err != nil {
-		return err
-	}
-
-	var authClient = &http.Client{
-		Timeout: time.Second * 60,
-	}
-	resp, err := authClient.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	t.SessionObject = resp.Cookies()
-	return nil
-}
-
-// getAuthRequest assembles the request to get the authenticated cookie
-func (t *CookieAuthTransport) buildAuthRequest() (*http.Request, error) {
-	body := struct {
-		Username string `json:"username"`
-		Password string `json:"password"`
-	}{
-		t.Username,
-		t.Password,
-	}
-
-	b := new(bytes.Buffer)
-	json.NewEncoder(b).Encode(body)
-
-	req, err := http.NewRequest("POST", t.AuthURL, b)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-	return req, nil
-}
-
-func (t *CookieAuthTransport) transport() http.RoundTripper {
-	if t.Transport != nil {
-		return t.Transport
-	}
-	return http.DefaultTransport
-}
-
-// JWTAuthTransport is an http.RoundTripper that authenticates all requests
-// using Jira's JWT based authentication.
-//
-// NOTE: this form of auth should be used by add-ons installed from the Atlassian marketplace.
-//
-// Jira docs: https://developer.atlassian.com/cloud/jira/platform/understanding-jwt
-// Examples in other languages:
-//
-//	https://bitbucket.org/atlassian/atlassian-jwt-ruby/src/d44a8e7a4649e4f23edaa784402655fda7c816ea/lib/atlassian/jwt.rb
-//	https://bitbucket.org/atlassian/atlassian-jwt-py/src/master/atlassian_jwt/url_utils.py
-type JWTAuthTransport struct {
-	Secret []byte
-	Issuer string
-
-	// Transport is the underlying HTTP transport to use when making requests.
-	// It will default to http.DefaultTransport if nil.
-	Transport http.RoundTripper
-}
-
-func (t *JWTAuthTransport) Client() *http.Client {
-	return &http.Client{Transport: t}
-}
-
-func (t *JWTAuthTransport) transport() http.RoundTripper {
-	if t.Transport != nil {
-		return t.Transport
-	}
-	return http.DefaultTransport
-}
-
-// RoundTrip adds the session object to the request.
-func (t *JWTAuthTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	req2 := cloneRequest(req) // per RoundTripper contract
-	exp := time.Duration(59) * time.Second
-	qsh := t.createQueryStringHash(req.Method, req2.URL)
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
-		"iss": t.Issuer,
-		"iat": time.Now().Unix(),
-		"exp": time.Now().Add(exp).Unix(),
-		"qsh": qsh,
-	})
-
-	jwtStr, err := token.SignedString(t.Secret)
-	if err != nil {
-		return nil, fmt.Errorf("jwtAuth: error signing JWT: %w", err)
-	}
-
-	req2.Header.Set("Authorization", fmt.Sprintf("JWT %s", jwtStr))
-	return t.transport().RoundTrip(req2)
-}
-
-func (t *JWTAuthTransport) createQueryStringHash(httpMethod string, jiraURL *url.URL) string {
-	canonicalRequest := t.canonicalizeRequest(httpMethod, jiraURL)
-	h := sha256.Sum256([]byte(canonicalRequest))
-	return hex.EncodeToString(h[:])
-}
-
-func (t *JWTAuthTransport) canonicalizeRequest(httpMethod string, jiraURL *url.URL) string {
-	path := "/" + strings.Replace(strings.Trim(jiraURL.Path, "/"), "&", "%26", -1)
-
-	var canonicalQueryString []string
-	for k, v := range jiraURL.Query() {
-		if k == "jwt" {
-			continue
-		}
-		param := url.QueryEscape(k)
-		value := url.QueryEscape(strings.Join(v, ""))
-		canonicalQueryString = append(canonicalQueryString, strings.Replace(strings.Join([]string{param, value}, "="), "+", "%20", -1))
-	}
-	sort.Strings(canonicalQueryString)
-	return fmt.Sprintf("%s&%s&%s", strings.ToUpper(httpMethod), path, strings.Join(canonicalQueryString, "&"))
-}
-
-// cloneRequest returns a clone of the provided *http.Request.
-// The clone is a shallow copy of the struct and its Header map.
-func cloneRequest(r *http.Request) *http.Request {
-	// shallow copy of the struct
-	r2 := new(http.Request)
-	*r2 = *r
-	// deep copy of the Header
-	r2.Header = make(http.Header, len(r.Header))
-	for k, s := range r.Header {
-		r2.Header[k] = append([]string(nil), s...)
-	}
-	return r2
 }
